+++ RJW_More_Genes/1.4/Source/RJW_More_Genes/CompAbilityEffect_PussyHeal.cs
@@ -95,6 +95,11 @@ namespace RJW_More_Genes
 				reason = "Rape is disabled";
 				return true;
 			}
+			else if (!RJWMGSettings.gene_pussyheal)
+			{
+				reason = "Disabled in modsettings";
+				return true;
+			}
 			return false;
         }