+++ RJW_More_Genes/1.4/Source/RJW_More_Genes/JobDefOf.cs
@@ -3,15 +3,14 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using RimWorld;
 using Verse;
+using RimWorld;
 
 namespace RJW_More_Genes
 {
     [DefOf]
-    public static class HediffDefOf
+    public static class JobDefOf
     {
-        [MayRequireRoyalty]
-        public static readonly HediffDef SexFrenzy;
+        public static readonly JobDef Mindbreak;
     }
 }